Changed: #825 Remove all warnings when compiling Ryzom

+++ code/nel/tools/3d/tile_edit_qt/browser_model.cpp
@@ -691,7 +691,7 @@ tilelist::iterator TileList::Get(int i, int n)
 
 int TileList::GetSize(int n)
 {
-	return theList[n].size();
+	return (int)theList[n].size();
 }
